+++ eslint.config.mjs
@@ -1,0 +1,10 @@
+// @ts-check
+import { defineConfig, globalIgnores } from "eslint/config";
+import eslint from '@eslint/js';
+import tseslint from 'typescript-eslint';
+
+export default defineConfig(
+    globalIgnores(["**/out", "**/dist", "**/*.d.ts"]),
+    eslint.configs.recommended,
+    tseslint.configs.recommended,
+);